% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{plot_partial_pooling}
\alias{plot_partial_pooling}
\title{Plot partial pooling estimates for a linear mixed effects model}
\usage{
plot_partial_pooling(df, y_var, x_var, group, ...)
}
\arguments{
\item{df}{a data.frame containing the variables used in the model}

\item{y_var}{the raw (unquoted) name of the dependent variable}

\item{x_var}{the raw (unquoted) name of the independent variable}

\item{group}{the raw (unquoted) name of the group factor}
}
\value{
a ggplot2 plot
}
\description{
Plot partial pooling estimates for a linear mixed effects model
}
\examples{
library(dplyr)
storms
plot_partial_pooling(storms, y_var = wind, x_var = pressure, group = year)
}
